refactor: implement layered architecture (separation of concerns)
