+++ src/utils/data_loader.py
@@ -1,128 +1,11 @@
 # utils/data_loader.py
-import json
-import os
-import threading
-import logging
-
-from utils.url_validator import is_valid_url, parse_resource
-
-DATA_FILE = os.path.join(os.path.dirname(__file__), "..", "..", "data", "projects.json")
-
-logger = logging.getLogger("devpath.data_loader")
-
-_projects_cache = None
-_cache_lock = threading.Lock()
-
-
-def validate_projects(projects):
-    """
-    Validate project dataset integrity.
-
-    Checks for:
-    - Duplicate project IDs
-    - Duplicate project titles (case-insensitive)
-    - Missing required fields
-    - Empty required string fields
-    - Malformed resource URLs (logs a warning, does not raise)
-
-    Raises:
-        ValueError: If any structural validation rule is violated.
-    """
-    seen_ids = set()
-    seen_titles = set()
-
-    required_fields = [
-        "id", "title", "skills", "level", "interest", "time",
-        "description", "features", "tech_stack", "roadmap",
-        "resources", "starter_code"
-    ]
-
-    for project in projects:
-        # Required fields
-        for field in required_fields:
-            if field not in project:
-                raise ValueError(f"Missing required field: {field}")
-
-            if isinstance(project[field], str) and not project[field].strip():
-                raise ValueError(
-                    f"Empty value for field '{field}' in project '{project.get('title', 'Unknown')}'"
-                )
-
-        # Duplicate IDs
-        project_id = project["id"]
-        if project_id in seen_ids:
-            raise ValueError(f"Duplicate project ID found: {project_id}")
-        seen_ids.add(project_id)
-
-        # Duplicate Titles
-        title = " ".join(project["title"].split()).lower()
-        if title in seen_titles:
-            raise ValueError(f"Duplicate project title found: {project['title']}")
-        seen_titles.add(title)
-
-        # Resource URL format validation — warn, do not raise
-        # Broken URLs in production are logged so they can be fixed in the
-        # data file without crashing the application.
-        for raw in project.get("resources", []):
-            parsed = parse_resource(raw)
-            url = parsed.get("url", "")
-            if url and not is_valid_url(url):
-                logger.warning(
-                    "Malformed resource URL in project '%s' (id=%s): %r",
-                    project.get("title", "Unknown"),
-                    project_id,
-                    url,
-                )
-
-
-def load_all_projects():
-    """Read and return the full list of projects from the JSON file.
-
-    Results are cached in memory after the first read so subsequent calls
-    do not hit the filesystem.
-    """
-    global _projects_cache
-    if _projects_cache is None:
-        with open(DATA_FILE, "r", encoding="utf-8") as f:
-            _projects_cache = json.load(f)
-        validate_projects(_projects_cache)
-    return _projects_cache
-
-
-def get_available_levels():
-    """Return all unique project levels."""
-    projects = load_all_projects()
-    return sorted({p["level"] for p in projects})
-
-
-def find_project_by_id(project_id):
-    """Return the project whose 'id' matches project_id, or None."""
-    for project in load_all_projects():
-        if project.get("id") == project_id:
-            return project
-    return None
-
-
-def get_project_stats():
-    """Return total_projects, unique_skills, and beginner_friendly counts."""
-    projects = load_all_projects()
-
-    all_skills = set()
-    beginner_friendly = 0
-    for p in projects:
-        for s in p.get("skills", []):
-            all_skills.add(s)
-        if p.get("level") == "Beginner":
-            beginner_friendly += 1
-
-    return {
-        "total_projects": len(projects),
-        "unique_skills": len(all_skills),
-        "beginner_friendly": beginner_friendly,
-    }
-
-def clear_cache():
-    """Reset the in-memory project cache (used in tests)."""
-    global _projects_cache
-    with _cache_lock:
-        _projects_cache = None
+# Compatibility shim forwarding to services.project_service and repositories.project_repository
+from repositories.project_repository import DATA_FILE
+from services.project_service import (
+    load_all_projects,
+    find_project_by_id,
+    get_available_levels,
+    get_project_stats,
+    clear_cache,
+    validate_projects,
+)
