+++ test_recommender.py
@@ -6,7 +6,7 @@ import os
 
 # Make sure imports resolve from the repo root regardless of where Python
 # looks by default.
-sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
+sys.path.insert(0, os.path.join(os.path.dirname(os.path.abspath(__file__)), "src"))
 
 from utils.recommender import (
     get_recommendations,
@@ -205,7 +205,7 @@ else:
 if isinstance(prog, list):
     passed(f"progression is a list  ({len(prog)} result(s))")
     for p in prog:
-        print(f"        → {p['project']['title']}  (gap_score: {p['gap_score']})")
+        print(f"        -> {p['project']['title']}  (gap_score: {p['gap_score']})")
 
 # ---------------------------------------------------------------------------
 # Summary
